// frontend/src/api/generated/models/healthFlagSeverity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type HealthFlagSeverity = typeof HealthFlagSeverity[keyof typeof HealthFlagSeverity];


export const HealthFlagSeverity = {
  Mild: 'Mild',
  Moderate: 'Moderate',
  Severe: 'Severe',
} as const;