// frontend/src/api/generated/models/healthFlagStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type HealthFlagStatus = typeof HealthFlagStatus[keyof typeof HealthFlagStatus];


export const HealthFlagStatus = {
  Active: 'Active',
  Improving: 'Improving',
  Resolved: 'Resolved',
} as const;