// frontend/src/api/generated/models/healthFlagType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type HealthFlagType = typeof HealthFlagType[keyof typeof HealthFlagType];


export const HealthFlagType = {
  Pain: 'Pain',
  Illness: 'Illness',
  Injury: 'Injury',
  Other: 'Other',
} as const;