// frontend/src/api/generated/models/painOrHealthFlagDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */
import type { HealthFlagSeverity } from './healthFlagSeverity';
import type { HealthFlagStatus } from './healthFlagStatus';
import type { HealthFlagType } from './healthFlagType';

export interface PainOrHealthFlagDto {
  id?: string;
  athleteUserId?: string;
  /** @nullable */
  relatedCheckInId?: string | null;
  type?: HealthFlagType;
  severity?: HealthFlagSeverity;
  status?: HealthFlagStatus;
  /** @nullable */
  bodyPart?: string | null;
  /** @nullable */
  description?: string | null;
  startedOnDate?: string;
  /** @nullable */
  resolvedOnDate?: string | null;
}